Clean up Plotter
The Plotter class did all its work in the constructor.  This triggered
a [misc-unused-raii] warning from clang-tidy when enabling the misc
checker.

Refactor Plotter by replacing the two constructors with two Plot
functions:
- rewrite the tests to use the Plot function
- Internally, the Plot functions instantiate a Plotter struct and use
  its members to plot the graphs.

This leads to a lot more calls from the Plotter and good usage.
Plotter is now more of a utility class for plotting graphviz.

Also, enable the clang-tidy misc checker as it now passes.