Clean up hicpp coding standards warnings
Fix warnings of the forms:
- hicpp-explicit-conversions
- hicpp-use-override

Explicitly do not fix hicpp-use-auto warnings Parser test but do
explain why.  In this test we want to enforce that Parser acts like an
iterator for the terms it generates.  It should have a
Parser::iterator type and begin and end should return values of that
type.  This test verifies that.

Enable hicpp checkers as we are now clean.